+++ HaselTweaks/Config/PluginConfig.cs
@@ -136,6 +136,7 @@ public class TweakConfigs
     public EnhancedMonsterNoteConfiguration EnhancedMonsterNote { get; init; } = new();
     public EnhancedTargetInfoConfiguration EnhancedTargetInfo { get; init; } = new();
     public ForcedCutsceneMusicConfiguration ForcedCutsceneMusic { get; init; } = new();
+    public GlamourDresserArmoireAlertConfiguration GlamourDresserArmoireAlert { get; init; } = new();
     public GearSetGridConfiguration GearSetGrid { get; init; } = new();
     public InventoryHighlightConfiguration InventoryHighlight { get; init; } = new();
     public LockWindowPositionConfiguration LockWindowPosition { get; init; } = new();
@@ -169,6 +170,7 @@ public static class PluginConfigExtension
         services.AddSingleton(pluginConfig.Tweaks.EnhancedMonsterNote);
         services.AddSingleton(pluginConfig.Tweaks.EnhancedTargetInfo);
         services.AddSingleton(pluginConfig.Tweaks.ForcedCutsceneMusic);
+        services.AddSingleton(pluginConfig.Tweaks.GlamourDresserArmoireAlert);
         services.AddSingleton(pluginConfig.Tweaks.GearSetGrid);
         services.AddSingleton(pluginConfig.Tweaks.InventoryHighlight);
         services.AddSingleton(pluginConfig.Tweaks.LockWindowPosition);